Optimize build configuration and enhance tensor operations
- Increased optimization level in the release profile for better performance.
- Enabled link-time optimization and adjusted code generation settings for improved efficiency.
- Updated README to reflect a more user-friendly Discord badge.
- Refactored shape manipulation functions to ensure index checks use `usize` for better type safety.
- Improved pointer arithmetic in tensor operations for clarity and performance.
- Enhanced error handling in tensor operations to provide more informative messages.
- Added new utility functions for multi-threaded processing of tensor operations.

+++ hpt-common/src/error/shape.rs
@@ -284,11 +284,11 @@ impl ShapeError {
 
     /// Check if the index is out of range
     #[track_caller]
-    pub fn check_index_out_of_range(index: i64, dim: i64) -> Result<(), Self> {
+    pub fn check_index_out_of_range(index: usize, dim: usize) -> Result<(), Self> {
         if index >= dim || index < 0 {
             return Err(Self::DimOutOfRange {
-                expected: 0..dim,
-                actual: index,
+                expected: 0..dim as i64,
+                actual: index as i64,
                 location: Location::caller(),
             });
         }
